requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
alembic
pyJWT
bcrypt
passlib

